+++ Plugins/TableProPluginKit/ConnectionMode.swift
@@ -1,0 +1,9 @@
+//
+//  ConnectionMode.swift
+//  TableProPluginKit
+//
+
+public enum ConnectionMode: String, Codable, Sendable {
+    case network
+    case fileBased
+}